requirements.txt
python-dotenv~=1.0.0
flask~=2.3.3
langchain~=0.0.277
PyJWT~=2.8.0
requests~=2.31.0